moving icons.js to root folder, creating new file RecommendedIcon.js for the Dimension component
