+++ packages/app/src/components/Dimensions/DimensionOptions.js
@@ -4,7 +4,7 @@ import { connect } from 'react-redux';
 import i18n from '@dhis2/d2-i18n';
 import MenuItem from '@material-ui/core/MenuItem';
 import DropDown from './DropDown';
-import { MoreHorizontal } from './icons';
+import { MoreHorizontal } from '../../icons';
 import { acAddUiLayoutDimensions } from '../../actions/ui';
 
 const style = {
